Separate GoogleEvent logic into appropriate modules
