+++ app/controllers/events_controller.rb
@@ -8,7 +8,7 @@ class EventsController < ApplicationController
     render json: {error: 'Google Server error'}, status: :service_unavailable
   end
 
-  rescue_from Google::Apis::ClientError, GoogleEvent::InvalidParamsError do |exception|
+  rescue_from Google::Apis::ClientError, GoogleCalendar::Adding::InvalidParamsError do |exception|
     error_data = {error: exception.message}
     case params[:action]
     when 'create'
@@ -25,7 +25,7 @@ class EventsController < ApplicationController
     render json: {error: 'Authorization error'}, status: :unauthorized
   end
 
-  rescue_from GoogleEvent::EventInTimeSpanError do |message|
+  rescue_from GoogleCalendar::Adding::EventInTimeSpanError do |message|
     render json: {conference_room_id: [message]}, status: :unprocessable_entity
   end
 
@@ -35,14 +35,14 @@ class EventsController < ApplicationController
 
   def create
     conference_room_id = event_params[:conference_room_id]
-    google_event_params = GoogleEvent.process_params(event_params)
-    data = GoogleEvent.create(session[:credentials], conference_room_id, google_event_params)
+    google_event_params = GoogleCalendar::GoogleEvent.process_params(event_params)
+    data = GoogleCalendar::GoogleEvent.create(session[:credentials], conference_room_id, google_event_params)
     render json: data.to_json, status: :created
   end
 
   def destroy
     event_id = params[:id]
-    GoogleEvent.delete(session[:credentials], event_id)
+    GoogleCalendar::GoogleEvent.delete(session[:credentials], event_id)
     redirect_to root_path
   end
