+++ app/models/google_calendar/client.rb
@@ -1,0 +1,11 @@
+module GoogleCalendar
+  module Client
+    def calendar_service(credentials)
+      Google::Apis::CalendarV3::CalendarService.new.tap { |s| s.authorization = client(credentials) }
+    end
+
+    def client(credentials)
+      Signet::OAuth2::Client.new(JSON.parse(credentials))
+    end
+  end
+end
