+++ spec/models/google_calendar/listing_spec.rb
@@ -1,6 +1,10 @@
 require 'rails_helper'
 
-describe GoogleEvent do
+class Listing
+  extend GoogleCalendar::Listing
+end
+
+describe Listing do
   describe '.list_events' do
     let!(:sample_conference_room1) { create :conference_room }
     let!(:sample_conference_room2) { create :conference_room }
@@ -21,7 +25,7 @@ describe GoogleEvent do
                                           end: Google::Apis::CalendarV3::EventDateTime.new(date_time: end_time2),
                                           summary: 'Event2', description: sample_summary,
                                           attendees: [Google::Apis::CalendarV3::EventAttendee.new(self: true, response_status: 'accepted')]
-                                         )
+      )
     end
 
     let(:sample_events) do
@@ -60,7 +64,7 @@ describe GoogleEvent do
 
       context 'room rejected event' do
         before do
-          google_event1.attendees.first.response_status = GoogleEvent.singleton_class::GOOGLE_EVENT_DECLINED_RESPONSE
+          google_event1.attendees.first.response_status = described_class.singleton_class::GOOGLE_EVENT_DECLINED_RESPONSE
         end
         it 'ignores event' do
           expect(described_class.list_events('', sample_email, start_time1, start_time2)).to satisfy do |response|
@@ -72,7 +76,7 @@ describe GoogleEvent do
         before do
           google_event1.attendees.push(
             Google::Apis::CalendarV3::EventAttendee.new(
-              response_status: GoogleEvent.singleton_class::GOOGLE_EVENT_DECLINED_RESPONSE
+              response_status: described_class.singleton_class::GOOGLE_EVENT_DECLINED_RESPONSE
             )
           )
         end
@@ -137,135 +141,6 @@ describe GoogleEvent do
     end
   end
 
-  describe '.create' do
-    context 'given invalid params' do
-      let(:invalid_event_response) { [false, {start: ['is missing'], end: ['is missing']}] }
-      it 'raises GoogleEvent::InvalidParamsException' do
-        expect { described_class.create({}, 0, {}) }.to raise_error(GoogleEvent::InvalidParamsError)
-      end
-    end
-
-    context 'valid params' do
-      context 'other events exitis' do
-        let(:credentials) { :credentials }
-        let(:first_event) { double('Event', summary: 'Summary') }
-        let(:second_event) { double('Event', summary: 'Meeting') }
-        let(:start_time) { Time.now }
-        let(:end_time) { Time.now + 3.hour }
-        let(:event_data) do
-          {
-            attendees: [],
-            start: {date_time: start_time},
-            end: {date_time: end_time}
-          }
-        end
-        let!(:room) { create :conference_room }
-
-        before do
-          allow(described_class).to receive(:events_in_span) do
-            double('EventList', items: [first_event, second_event])
-          end
-        end
-
-        it 'raises EventInTimeSpanError' do
-          expect do
-            described_class.create(credentials,
-                                   room.id,
-                                   start: {date_time: start_time},
-                                   end: {date_time: end_time})
-          end.to raise_error(
-            GoogleEvent::EventInTimeSpanError,
-            'Already 2 events in time span(Summary, Meeting).'
-          )
-        end
-      end
-      context 'no other events' do
-        let(:credentials) { :credentials }
-        let(:service) { double(:calendar_service) }
-        let(:start_time) { Time.now }
-        let(:end_time) { Time.now + 3.hour }
-        let(:event_data) do
-          {
-            attendees: [],
-            start: {date_time: start_time},
-            end: {date_time: end_time}
-          }
-        end
-        let!(:room) { create :conference_room }
-
-        before do
-          allow(described_class).to receive(:events_in_span) do
-            double('EventList', items: [])
-          end
-          allow(described_class).to receive(:calendar_service) { service }
-          allow(service).to receive(:insert_event) { true }
-        end
-        it 'creates event' do
-          expect(service).to receive(:insert_event).with(
-            'primary',
-            Google::Apis::CalendarV3::Event
-          )
-          described_class.create(credentials,
-                                 room.id,
-                                 start: {date_time: start_time},
-                                 end: {date_time: end_time})
-        end
-      end
-    end
-  end
-
-  describe 'EVENT_SCHEMA' do
-    let(:schema) { GoogleEvent::EVENT_SCHEMA }
-    let(:time1) { Faker::Time.forward 5 }
-    let(:time2) { Faker::Time.forward 5 }
-    let(:params) { {start: {date_time: time1}, end: {date_time: time2}} }
-    context 'valid' do
-      it { expect(schema.call(params).success?).to eq true }
-    end
-    context 'no start' do
-      it 'is invalid' do
-        expect(schema.call(end: {date_time: 'asdf'}).messages[:start]).to be_present
-      end
-    end
-    context 'no end' do
-      it 'is invalid' do
-        expect(schema.call(start: {date_time: 'asdf'}).messages[:end]).to be_present
-      end
-    end
-
-    context 'no end datetime' do
-      it 'is invalid' do
-        expect(schema.call(start: {date_time: time1}, end: {test: nil}).messages[:end]).to be_present
-      end
-    end
-    context 'no start datetime' do
-      it 'is invalid' do
-        expect(schema.call(end: {date_time: time1}, start: {test: nil}).messages[:start]).to be_present
-      end
-    end
-  end
-
-  describe '.add_rooms_to_event' do
-    context 'given array of calendar room ids' do
-      let(:mordor_email) { 'u2i.com_2d3631343934393033313035@resource.calendar.google.com' }
-      let(:neverland_email) { 'u2i.com_3530363130383730383638@resource.calendar.google.com' }
-      let(:expected_result) do
-        {
-          attendees: [
-            {email: mordor_email}
-          ],
-          location: first_room.title
-        }
-      end
-      let(:params) { {} }
-      let!(:first_room) { create(:conference_room, email: mordor_email) }
-      it 'adds new key in hash and assigns array of conference room emails to it' do
-        described_class.add_room_to_event(params, first_room.id)
-        expect(params).to eq expected_result
-      end
-    end
-  end
-
   describe '.mark_user_events' do
     let(:credentials) { {} }
     let(:user_email) { 'user@example.com' }
@@ -276,28 +151,9 @@ describe GoogleEvent do
 
     it 'sets event[:creator][:self] flag to true if user is creator of the event' do
       allow(GoogleOauth).to receive(:user_email) { user_email }
-      GoogleEvent.mark_user_events(user_email, all_events)
+      described_class.mark_user_events(user_email, all_events)
       expect(event1[:creator][:self]).to eq true
       expect(event2[:creator][:self]).to eq false
-    end
-  end
-  describe '.event_in_span' do
-    let(:service) { double(:calendar_service) }
-    let(:credentials) { :credentials }
-    let(:conference_room) { {email: 'email@sample.com'.freeze, key: :value} }
-    let(:start_time) { Time.now }
-    let(:end_time) { Time.now + 3.hour }
-
-    before do
-      allow(described_class).to receive(:calendar_service) { service }
-    end
-
-    it 'calls calendar_service' do
-      expect(described_class).to receive(:calendar_service).with(credentials)
-      expect(service).to receive(:list_events).with(conference_room[:email],
-                                                    time_min: start_time,
-                                                    time_max: end_time)
-      described_class.events_in_span(credentials, conference_room, start_time, end_time)
     end
   end
 end
