+++ spec/requests/events_spec.rb
@@ -11,8 +11,8 @@ RSpec.describe 'Events', type: :request do
     context 'given invalid event attributes' do
       let(:exception) { Google::Apis::ClientError.new('error') }
       it 'responds with 422' do
-        allow(GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
-        allow(GoogleEvent).to receive(:process_params) { |arg| arg }
+        allow(GoogleCalendar::GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
+        allow(GoogleCalendar::GoogleEvent).to receive(:process_params) { |arg| arg }
         allow_any_instance_of(EventsController).to receive(:event_params) { {conference_room_id: 2} }
         allow_any_instance_of(EventsController).to receive(:session) { {credentials: 123} }
         post events_path
@@ -23,8 +23,8 @@ RSpec.describe 'Events', type: :request do
     context 'google server does not respond' do
       let(:exception) { Google::Apis::ServerError.new('error') }
       it 'responds with 503' do
-        allow(GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
-        allow(GoogleEvent).to receive(:process_params) { |arg| arg }
+        allow(GoogleCalendar::GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
+        allow(GoogleCalendar::GoogleEvent).to receive(:process_params) { |arg| arg }
         allow_any_instance_of(EventsController).to receive(:event_params) { {conference_room_id: 2} }
         allow_any_instance_of(EventsController).to receive(:session) { {credentials: 123} }
         post events_path
@@ -35,8 +35,8 @@ RSpec.describe 'Events', type: :request do
     context 'user is not authorized' do
       let(:exception) { Google::Apis::AuthorizationError.new('error') }
       it 'responds with 401' do
-        allow(GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
-        allow(GoogleEvent).to receive(:process_params) { |arg| arg }
+        allow(GoogleCalendar::GoogleEvent).to receive(:create).with(any_args).and_raise(exception)
+        allow(GoogleCalendar::GoogleEvent).to receive(:process_params) { |arg| arg }
         allow_any_instance_of(EventsController).to receive(:event_params) { {conference_room_id: 2} }
         allow_any_instance_of(EventsController).to receive(:session) { {credentials: 123} }
         post events_path
@@ -46,8 +46,8 @@ RSpec.describe 'Events', type: :request do
 
     context 'successfully added new event' do
       it 'repond with 200' do
-        allow(GoogleEvent).to receive(:create).with(any_args) { {} }
-        allow(GoogleEvent).to receive(:process_params) { |arg| arg }
+        allow(GoogleCalendar::GoogleEvent).to receive(:create).with(any_args) { {} }
+        allow(GoogleCalendar::GoogleEvent).to receive(:process_params) { |arg| arg }
         allow_any_instance_of(EventsController).to receive(:event_params) { {conference_room_id: 2} }
         allow_any_instance_of(EventsController).to receive(:session) { {credentials: 123} }
         post events_path
@@ -62,7 +62,7 @@ RSpec.describe 'Events', type: :request do
     context 'request is forbidden' do
       let(:exception) { Google::Apis::ClientError.new('forbidden error') }
       it 'responds with 403' do
-        allow(GoogleEvent).to receive(:delete).and_raise(exception)
+        allow(GoogleCalendar::GoogleEvent).to receive(:delete).and_raise(exception)
         allow_any_instance_of(EventsController).to receive(:session) { session }
         delete event_path event_id
         expect(response).to have_http_status :forbidden
@@ -70,7 +70,7 @@ RSpec.describe 'Events', type: :request do
     end
     context 'request is valid' do
       it 'redirects to root_path' do
-        allow(GoogleEvent).to receive(:delete) { true }
+        allow(GoogleCalendar::GoogleEvent).to receive(:delete) { true }
         allow_any_instance_of(EventsController).to receive(:session) { session }
         delete event_path event_id
         expect(response).to redirect_to root_path
